limit sudo timeout extension to linux only

+++ lib/functions.sh
@@ -30,17 +30,21 @@ load_recipe () {
 }
 
 extend_sudo_timeout () {
-  # reset sudo timeout
-  sudo -k
-  if [ "$RUN" == 1 ]; then
-    sudo -v
+  if is_linux; then
+    # reset sudo timeout
+    sudo -k
+    if [ "$RUN" == 1 ]; then
+      sudo -v
+    fi
+
+    log_info "extending sudo timeout"
+    execute_su "echo \"Defaults timestamp_timeout=1800\" > /etc/sudoers.d/expand_timeout"
   fi
-  
-  log_info "extending sudo timeout"
-  execute_su "echo \"Defaults timestamp_timeout=1800\" > /etc/sudoers.d/expand_timeout"
 }
 
 restore_sudo_timeout () {
-  log_info "restoring sudo timeout"
-  execute_su "rm /etc/sudoers.d/expand_timeout"
+  if is_linux; then
+    log_info "restoring sudo timeout"
+    execute_su "rm /etc/sudoers.d/expand_timeout"
+  fi
 }
